Make new plot legend & settings the default (#5287)
**User-Facing Changes**
Replace the old plot legend with the new legend and settings driven
configuration.

**Description**
Removes the old plot legend and replaces it with the new legend and
moves plot configuration into the settings sidebar.
