+++ packages/studio-base/src/AppSetting.ts
@@ -20,7 +20,6 @@ export enum AppSetting {
   // Experimental features
   SHOW_DEBUG_PANELS = "showDebugPanels",
   ENABLE_LEGACY_PLOT_PANEL = "enableLegacyPlotPanel",
-  ENABLE_PLOT_PANEL_SERIES_SETTINGS = "enablePlotPathSeriesSettings",
   ENABLE_NEW_TOPNAV = "enableNewTopNav",
 
   // Miscellaneous
