+++ packages/studio-base/src/panels/Plot/index.tsx
@@ -15,19 +15,18 @@ import DownloadIcon from "@mui/icons-material/Download";
 import { useTheme } from "@mui/material";
 import { compact, isNumber, uniq } from "lodash";
 import memoizeWeak from "memoize-weak";
-import { useEffect, useCallback, useMemo, ComponentProps, useState } from "react";
+import { ComponentProps, useCallback, useEffect, useMemo, useState } from "react";
 
 import { filterMap } from "@foxglove/den/collection";
 import { useShallowMemo } from "@foxglove/hooks";
 import {
-  Time,
   add as addTimes,
   fromSec,
   subtract as subtractTimes,
+  Time,
   toSec,
 } from "@foxglove/rostime";
 import { MessageEvent } from "@foxglove/studio";
-import { AppSetting } from "@foxglove/studio-base/AppSetting";
 import { useBlocksByTopic, useMessageReducer } from "@foxglove/studio-base/PanelAPI";
 import { MessageBlock } from "@foxglove/studio-base/PanelAPI/useBlocksByTopic";
 import parseRosPath, {
@@ -53,15 +52,13 @@ import {
   ChartDefaultView,
   TimeBasedChartTooltipData,
 } from "@foxglove/studio-base/components/TimeBasedChart";
-import { useAppConfigurationValue } from "@foxglove/studio-base/hooks";
-import { NewPlotLegend } from "@foxglove/studio-base/panels/Plot/NewPlotLegend";
 import { OnClickArg as OnChartClickArgs } from "@foxglove/studio-base/src/components/Chart";
 import { OpenSiblingPanel, PanelConfig, SaveConfig } from "@foxglove/studio-base/types/panels";
 import { PANEL_TITLE_CONFIG_KEY } from "@foxglove/studio-base/util/layout";
 import { getTimestampForMessage } from "@foxglove/studio-base/util/time";
 
 import PlotChart from "./PlotChart";
-import PlotLegend from "./PlotLegend";
+import { PlotLegend } from "./PlotLegend";
 import { downloadCSV } from "./csv";
 import { getDatasets } from "./datasets";
 import { DataSet, PlotDataByPath, PlotDataItem } from "./internalTypes";
@@ -506,10 +503,6 @@ function Plot(props: Props) {
 
   usePlotPanelSettings(config, saveConfig, focusedPath);
 
-  const [seriesSettings = false] = useAppConfigurationValue(
-    AppSetting.ENABLE_PLOT_PANEL_SERIES_SETTINGS,
-  );
-
   const stackDirection = useMemo(
     () => (legendDisplay === "top" ? "column" : "row"),
     [legendDisplay],
@@ -539,35 +532,18 @@ function Plot(props: Props) {
         fullWidth
         style={{ height: `calc(100% - ${PANEL_TOOLBAR_MIN_HEIGHT}px)` }}
       >
-        {seriesSettings === false && (
-          <PlotLegend
-            paths={yAxisPaths}
-            datasets={datasets}
-            currentTime={currentTimeSinceStart}
-            saveConfig={saveConfig}
-            showLegend={showLegend}
-            xAxisVal={xAxisVal}
-            xAxisPath={xAxisPath}
-            pathsWithMismatchedDataLengths={pathsWithMismatchedDataLengths}
-            legendDisplay={legendDisplay}
-            showPlotValuesInLegend={showPlotValuesInLegend}
-            sidebarDimension={sidebarDimension}
-          />
-        )}
-        {seriesSettings === true && (
-          <NewPlotLegend
-            onClickPath={(index: number) => setFocusedPath(["paths", String(index)])}
-            paths={yAxisPaths}
-            datasets={datasets}
-            currentTime={currentTimeSinceStart}
-            saveConfig={saveConfig}
-            showLegend={showLegend}
-            pathsWithMismatchedDataLengths={pathsWithMismatchedDataLengths}
-            legendDisplay={legendDisplay}
-            showPlotValuesInLegend={showPlotValuesInLegend}
-            sidebarDimension={sidebarDimension}
-          />
-        )}
+        <PlotLegend
+          paths={yAxisPaths}
+          datasets={datasets}
+          currentTime={currentTimeSinceStart}
+          onClickPath={(index: number) => setFocusedPath(["paths", String(index)])}
+          saveConfig={saveConfig}
+          showLegend={showLegend}
+          pathsWithMismatchedDataLengths={pathsWithMismatchedDataLengths}
+          legendDisplay={legendDisplay}
+          showPlotValuesInLegend={showPlotValuesInLegend}
+          sidebarDimension={sidebarDimension}
+        />
         <Stack flex="auto" alignItems="center" justifyContent="center" overflow="hidden">
           <PlotChart
             isSynced={xAxisVal === "timestamp" && isSynced}
